review model

+++ app/Models/Review.php
@@ -9,7 +9,7 @@ use Illuminate\Database\Eloquent\Relations\BelongsTo;
 class Review extends Model
 {
     protected $table = "reviews";
-    protected $keyType = 'bigint';
+    protected $keyType = 'int';
     protected $primaryKey = 'id';
     public $incrementing = true;
     public $timestamps = false;
